# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alpha Games API Docs",
        "description": "API for executing Alpha Games gaming e-commerce functionalities",
        "contact": {
            "name": "Felipe Silva",
            "email": "felipe.allware@gmail.com"
        },
        "license": {
            "name": "Licença API",
            "url": "https://github.com/felipesilva15/alpha-games-api/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-status": {
            "get": {
                "tags": [
                    "Order status"
                ],
                "summary": "List all order statuses",
                "operationId": "3dec5c4dfd643e6ee9c76656d763861c",
                "parameters": [
                    {
                        "name": "STATUS_DESC",
                        "in": "query",
                        "description": "Order status description",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order status list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderStatus"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order status"
                ],
                "summary": "Registers an order status",
                "operationId": "0fb8bb32d4c6b31e7c37ffeb9bcaba5c",
                "requestBody": {
                    "description": "Data for creating a new order status",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "STATUS_DESC": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered order status data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "STATUS_ID": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "STATUS_DESC": {
                                            "type": "string",
                                            "example": "Em andamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order-status/{id}": {
            "get": {
                "tags": [
                    "Order status"
                ],
                "summary": "List an order status by ID",
                "operationId": "afeddd47a5f76adc17ce6e7d619b98a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order Status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order status data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "STATUS_ID": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "STATUS_DESC": {
                                            "type": "string",
                                            "example": "Em andamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Order status"
                ],
                "summary": "Update an order status",
                "operationId": "2cd7ab3d82b8ad4905149cb56cd60a43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for update order status",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "STATUS_DESC": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated order status data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "STATUS_ID": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "STATUS_DESC": {
                                            "type": "string",
                                            "example": "Em andamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "RRecord not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Order status"
                ],
                "summary": "Deletes an order status",
                "operationId": "1211505793d3368926d238d3d941850b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro deletado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cep/{cep}": {
            "get": {
                "tags": [
                    "Search CEP"
                ],
                "summary": "Get address by CEP",
                "operationId": "fdea74b320e985d49572baaf52c262de",
                "parameters": [
                    {
                        "name": "cep",
                        "in": "path",
                        "description": "CEP",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid CEP",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "External tool error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AddressDTO": {
                "properties": {
                    "cep": {
                        "type": "string",
                        "example": "01001000"
                    },
                    "logradouro": {
                        "type": "string",
                        "example": "Praça da Sé"
                    },
                    "bairro": {
                        "type": "string",
                        "example": "Sé"
                    },
                    "cidade": {
                        "type": "string",
                        "example": "São Paulo"
                    },
                    "uf": {
                        "type": "string",
                        "example": "SP"
                    }
                },
                "type": "object"
            },
            "ApiError": {
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "EXCPHAND001"
                    },
                    "endpoint": {
                        "type": "string",
                        "example": "api/endpoint"
                    },
                    "message": {
                        "type": "string",
                        "example": "Ocorreu um erro."
                    }
                },
                "type": "object"
            },
...
            "OrderStatus": {
                "properties": {
                    "STATUS_ID": {
                        "type": "integer",
                        "example": 1
                    },
                    "STATUS_DESC": {
                        "type": "string",
                        "example": "Em andamento"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}